Revert "Move query ID generation from `KademliaHandle` to `Kademlia`"
This reverts commit 982c73dcf0b02ba03febe4e892d69411bf41a207.
