+++ src/protocol/libp2p/kademlia/mod.rs
@@ -166,9 +166,6 @@ pub(crate) struct Kademlia {
 
     /// Query executor.
     executor: QueryExecutor,
-
-    /// Next query ID.
-    next_query_id: usize,
 }
 
 impl Kademlia {
@@ -210,7 +207,6 @@ impl Kademlia {
             provider_ttl: config.provider_ttl,
             replication_factor: config.replication_factor,
             engine: QueryEngine::new(local_peer_id, config.replication_factor, PARALLELISM_FACTOR),
-            next_query_id: 0usize,
         }
     }
 
@@ -945,10 +941,7 @@ impl Kademlia {
                 },
                 command = self.cmd_rx.recv() => {
                     match command {
-                        Some(KademliaCommand::FindNode { peer, query_id_tx }) => {
-                            let query_id = self.next_query_id();
-                            let _ = query_id_tx.send(query_id);
-
+                        Some(KademliaCommand::FindNode { peer, query_id }) => {
                             tracing::debug!(
                                 target: LOG_TARGET,
                                 ?peer,
@@ -964,10 +957,7 @@ impl Kademlia {
                                     .into()
                             );
                         }
-                        Some(KademliaCommand::PutRecord { mut record, query_id_tx }) => {
-                            let query_id = self.next_query_id();
-                            let _ = query_id_tx.send(query_id);
-
+                        Some(KademliaCommand::PutRecord { mut record, query_id }) => {
                             tracing::debug!(
                                 target: LOG_TARGET,
                                 query = ?query_id,
@@ -995,13 +985,10 @@ impl Kademlia {
                         }
                         Some(KademliaCommand::PutRecordToPeers {
                             mut record,
-                            query_id_tx,
+                            query_id,
                             peers,
                             update_local_store,
                         }) => {
-                            let query_id = self.next_query_id();
-                            let _ = query_id_tx.send(query_id);
-
                             tracing::debug!(
                                 target: LOG_TARGET,
                                 query = ?query_id,
@@ -1041,11 +1028,8 @@ impl Kademlia {
                         Some(KademliaCommand::StartProviding {
                             key,
                             public_addresses,
-                            query_id_tx
+                            query_id
                         }) => {
-                            let query_id = self.next_query_id();
-                            let _ = query_id_tx.send(query_id);
-
                             tracing::debug!(
                                 target: LOG_TARGET,
                                 query = ?query_id,
@@ -1071,10 +1055,7 @@ impl Kademlia {
                                     .into(),
                             );
                         }
-                        Some(KademliaCommand::GetRecord { key, quorum, query_id_tx }) => {
-                            let query_id = self.next_query_id();
-                            let _ = query_id_tx.send(query_id);
-
+                        Some(KademliaCommand::GetRecord { key, quorum, query_id }) => {
                             tracing::debug!(target: LOG_TARGET, ?key, "get record from DHT");
 
                             match (self.store.get(&key), quorum) {
@@ -1156,14 +1137,6 @@ impl Kademlia {
                 }
             }
         }
-    }
-
-    /// Allocate next query ID.
-    fn next_query_id(&mut self) -> QueryId {
-        let query_id = self.next_query_id;
-        self.next_query_id += 1;
-
-        QueryId(query_id)
     }
 }
